+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Tidarr notable changes.
 - [Front] Add Tiddl TOML config live editor
 ### 🖍 Changed
 - [Front] Update some config dialog UI
+- [API] Restore mix-to-playlist processing to get more metadatas
 ### 🐛 Fixed
 - [Front] Fix `video_quality` front error